agent_with_MCP/requirements.txt
fastmcp
google-adk
requests
